Move "Digest" related stuff into its own file
