+++ clapi-hs.cabal
@@ -40,6 +40,7 @@ library
                      , Clapi.Types.AssocList
                      , Clapi.Types.Base
                      , Clapi.Types.Definitions
+                     , Clapi.Types.Digests
                      , Clapi.Types.Dkmap
                      , Clapi.Types.Messages
                      , Clapi.Types.Path
